+++ DataAccessLayer/Serialisation/Serialiser.java
@@ -1,6 +1,8 @@
-package DataAccessLayer;
+package DataAccessLayer.Serialisation;
 
 import java.io.*;
+
+import DataAccessLayer.user;
 
 public class Serialiser {
     public void Serialise(user user){
